requirements.txt
# MySQL и ORM
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0
alembic>=1.13.0

# Google Sheets
gspread>=6.0.0
google-auth>=2.25.0

# T-Bank API
httpx>=0.27.0

# Telegram
python-telegram-bot>=21.0

# Генерация PDF (акт)
reportlab>=4.0.0

# Конфигурация
python-dotenv>=1.0.0
